test(storage): extract shared backend conflict tracker tests
Move ConflictTracker/RecordGuard unit tests into shared_tests.rs.
Preserve the all(test, not(wasm32)) gate and module path
backends::shared::tests.
